Remove fetch functionality, add Router
The FetchService was apparently removed from yew since the version
available in crates.io (or on the documents). I'm assuming we're to
use js-sys/web-sys for this functionality, and that will be the
next job.

For now, mock out the Router to make sure that we can navigate
between views.