Fix build issue caused by multiple-version dependency problem
Also, fix a panic that occurred due to the incorrect implementation
of Deserialize for Option<NaiveDateTime>. chronotype/chrono
includes an implementation for this on the HEAD of the master branch
(upcoming in 4.20 release), so update to that version. Unfortunately,
the HEAD of master also has a build problem for
wasm32-unknown-unknown, so resolve the build problem in a fork
(and create a PR), and use the forked version until merge.