Begin migrating from Diesel to sqlx
I have grown tired of finagling with tracking master of Diesel,
so I'm migrating the application to using sqlx instead. This commit
marks the beginning of the transition, and also cleans up the
model crate to remove any binding dependencies that would tie the
application model to a particular dependency, environment, or
implementation.