Remove negative numbers from the model
Previously, amounts of money could be specified as positive integers
or negative integers (where the numbers represented a monetary
value in cents). This became confusing, so now update the model
only to allow positive integers. This constraint is only currently
enforced by the algorithms and (in the future) form validation.